Refactored building environment.

+++ scil/code/scil-internal.h
@@ -12,7 +12,7 @@ typedef struct{
     int (*compress)(const scil_context* ctx, byte* restrict compressed_buf_out, size_t* restrict out_size, const double*restrict data_in, const size_t in_size);
     int (*decompress)(const scil_context* ctx, double*restrict data_out, size_t*restrict out_size, const byte*restrict compressed_buf_in, const size_t in_size);
     const char * name;
-    int magic_number;
+    byte magic_number;
 } scil_compression_algorithm;
